@@ -59,6 +59,6 @@
     "types": "*"
   },
   "engines": {
-    "node": ">=18.17.0"
+    "node": ">=20"
   }
 }
