@@ -51,6 +51,6 @@
     "types": "*"
   },
   "engines": {
-    "node": ">=18.17.0"
+    "node": ">=20"
   }
 }
