@@ -39,10 +39,11 @@ Create the file \`src/modules/my-fulfillment/index.ts\` with the following conte
 
 \`\`\`ts title="src/modules/my-fulfillment/index.ts"
 import MyFulfillmentProviderService from "./service"
+import { Module } from "@medusajs/utils"
 
-export default {
+export default Module("my-fulfillment", {
   service: MyFulfillmentProviderService,
-}
+})
 \`\`\`
 
 This exports the module's definition, indicating that the \`MyFulfillmentProviderService\` is the main service of the module.`,
@@ -51,7 +52,7 @@ This exports the module's definition, indicating that the \`MyFulfillmentProvide
 To use your Fulfillment Provider Module, add it to the \`providers\` array of the Fulfillment Module:
 
 \`\`\`js title="medusa-config.js"
-const { Modules } = require("@medusajs/modules-sdk")
+import { Modules } from "@medusajs/utils"
 
 // ...
 
@@ -64,12 +65,9 @@ module.exports = defineConfig({
         providers: [
           {
             resolve: "./modules/my-fulfillment",
+            id: "my-fulfillment",
             options: {
-              config: {
-                "my-fulfillment": {
-                  // provider options...
-                },
-              },
+              // provider options...
             },
           },
         ],
