@@ -18,11 +18,12 @@ const TagsOperationDescriptionSectionResponses = ({
         className={clsx("[&>details:not(:first-of-type)>summary]:border-t-0")}
       >
         {Object.entries(responses).map(([code, response], index) => {
+          const successCode = code.startsWith("20")
           return (
             <Fragment key={index}>
               {response.content && (
                 <>
-                  {(code === "200" || code === "201") && (
+                  {successCode && (
                     <>
                       <DetailsSummary
                         title={`${code} ${response.description}`}
@@ -43,7 +44,7 @@ const TagsOperationDescriptionSectionResponses = ({
                       />
                     </>
                   )}
-                  {code !== "200" && code !== "201" && (
+                  {!successCode && (
                     <Details
                       summaryElm={
                         <DetailsSummary
@@ -72,12 +73,8 @@ const TagsOperationDescriptionSectionResponses = ({
                   title={`${code} ${response.description}`}
                   subtitle={"Empty response"}
                   badge={
-                    <Badge
-                      variant={
-                        code === "200" || code === "201" ? "green" : "red"
-                      }
-                    >
-                      {code === "200" || code === "201" ? "Success" : "Error"}
+                    <Badge variant={successCode ? "green" : "red"}>
+                      {successCode ? "Success" : "Error"}
                     </Badge>
                   }
                   expandable={false}
