@@ -17,8 +17,8 @@ export const sidebar = sidebarAttachHrefCommonOptions(
       title: "The Basics",
       children: [
         {
-          path: "/basics/important-directories-files",
-          title: "Important Directories and Files",
+          path: "/basics/project-directories-files",
+          title: "Project Directories and Files",
         },
         {
           path: "/basics/medusa-container",
@@ -86,12 +86,12 @@ export const sidebar = sidebarAttachHrefCommonOptions(
               title: "Middlewares",
             },
             {
-              path: "/advanced-development/api-routes/cors",
-              title: "Handling CORS",
-            },
-            {
               path: "/advanced-development/api-routes/protected-routes",
               title: "Protected Routes",
+            },
+            {
+              path: "/advanced-development/api-routes/cors",
+              title: "Handling CORS",
             },
           ],
         },
@@ -111,16 +111,16 @@ export const sidebar = sidebarAttachHrefCommonOptions(
               title: "Module Isolation",
             },
             {
-              path: "/advanced-development/modules/remote-query",
-              title: "Remote Query",
-            },
-            {
-              path: "/advanced-development/modules/link-modules",
-              title: "Link Modules",
+              path: "/advanced-development/modules/module-links",
+              title: "Module Links",
             },
             {
               path: "/advanced-development/modules/remote-link",
               title: "Remote Link",
+            },
+            {
+              path: "/advanced-development/modules/remote-query",
+              title: "Remote Query",
             },
             {
               path: "/advanced-development/modules/options",
@@ -129,6 +129,7 @@ export const sidebar = sidebarAttachHrefCommonOptions(
           ],
         },
         {
+          path: "/advanced-development/data-models",
           title: "Data Models",
           children: [
             {
@@ -158,15 +159,6 @@ export const sidebar = sidebarAttachHrefCommonOptions(
             {
               path: "/advanced-development/data-models/searchable-property",
               title: "Searchable Property",
-            },
-          ],
-        },
-        {
-          title: "Loaders",
-          children: [
-            {
-              path: "/advanced-development/loaders/outside-modules",
-              title: "Create Outside Module",
             },
           ],
         },
@@ -260,7 +252,7 @@ export const sidebar = sidebarAttachHrefCommonOptions(
       ],
     },
     {
-      path: "/architectural-concepts/architectural-modules",
+      path: "/architectural-modules",
       title: "Architectural Modules",
     },
     {
@@ -274,10 +266,6 @@ export const sidebar = sidebarAttachHrefCommonOptions(
         {
           path: "/debugging-and-testing/tools",
           title: "Tools",
-        },
-        {
-          path: "/debugging-and-testing/feature-flags",
-          title: "Feature Flags",
         },
       ],
     },
